+++ zarrs_codec/src/lib.rs
@@ -36,7 +36,7 @@ pub use codec_traits::array_to_bytes::{
 };
 pub use codec_traits::bytes_partial_sync::{BytesPartialDecoderTraits, BytesPartialEncoderTraits};
 pub use codec_traits::bytes_to_bytes::BytesToBytesCodecTraits;
-pub use codec_traits::subchunking::EncodedSubchunk;
+pub use codec_traits::subchunking::{EncodedSubchunk, enclosing_chunk_indices};
 pub use codec_traits::{CodecTraits, CodecTraitsV2, CodecTraitsV3};
 
 #[cfg(feature = "async")]